refactor: Separa os componentes da página inicial em partes menores para serem reutilizadas nas próximas páginas.
