+++ frontend/src/style/MainPage.css
@@ -7,109 +7,6 @@
   background-color: #fff;
 }
 
-.navbar {
-  background-color: #fff; /* Ajuste conforme necessário */
-  color: white;
-  padding: 40px;
-  text-align: center;
-  position: relative;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2); /* Sombra no canto inferior */
-}
-
-.navbar .user-icon,
-.navbar .left-icon {
-  position: absolute;
-  top: 50%; /* Centraliza verticalmente no meio da altura do navbar */
-  transform: translateY(-50%); /* Ajusta o alinhamento vertical ao meio */
-}
-
-.navbar .user-icon {
-  right: 20px; /* Ajuste para posicionar mais para a direita */
-  width: 50px; /* Ajuste o tamanho desejado para o ícone do usuário */
-  height: 50px; /* Ajuste o tamanho desejado para o ícone do usuário */
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 50%; /* Faz o ícone ficar redondo */
-  color: white;
-  font-weight: bold;
-  font-size: 24px; /* Ajuste conforme necessário */
-  text-transform: uppercase;
-  cursor: pointer; /* Adiciona o cursor de ponteiro para indicar que é clicável */
-}
-
-.navbar .left-icon {
-  left: 10px;
-  width: 160px;  /* Ajuste conforme necessário */
-  height: auto; /* Mantém a proporção da altura com base na largura */
-  padding-top: 5px;
-  padding-left: 20px;
-}
-
-.navbar img {
-  width: 100%; /* A imagem ocupa 100% do tamanho do contêiner */
-  height: auto; /* Mantém a proporção da altura com base na largura */
-  object-fit: contain; /* Ajusta a imagem para se encaixar no contêiner, mantendo a proporção */
-  border: none; /* Remove qualquer borda */
-  border-radius: 0; /* Remove arredondamento */
-}
-
-.user-box {
-  position: absolute;
-  top: 80px; /* Ajusta a posição logo abaixo do ícone do usuário */
-  right: 20px; /* Ajuste conforme necessário */
-  background-color: #fff;
-  border: 1px solid #ccc;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  border-radius: 8px;
-  padding: 20px;
-  text-align: center;
-  z-index: 10; /* Garante que a caixa apareça sobre outros elementos */
-  width: 200px; /* Ajuste a largura conforme necessário */
-}
-
-.user-box-icon {
-  width: 80px; /* Tamanho maior para o ícone dentro da caixa */
-  height: 80px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 50%;
-  color: white;
-  font-weight: bold;
-  font-size: 36px;
-  text-transform: uppercase;
-  margin: 0 auto 10px auto; /* Centraliza o ícone e adiciona margem inferior */
-}
-
-.user-box-name {
-  font-size: 18px;
-  font-weight: bold;
-  margin-bottom: 5px;
-}
-
-.user-box-type {
-  font-size: 14px;
-  margin-bottom: 15px;
-}
-
-.user-box-button {
-  display: block;
-  width: 100%;
-  padding: 10px;
-  margin-bottom: 10px;
-  border: 1px solid #ccc;
-  border-radius: 20px;
-  background-color: #fff;
-  color: rgb(0, 0, 0);
-  cursor: pointer;
-}
-
-.user-box-button:hover {
-  background-color: #DB1E2F; /* Cor de fundo ao passar o mouse */
-  color: #fff
-}
-
 .content {
   flex: 1;
   display: flex;
@@ -125,19 +22,4 @@
   width: 400px;
   text-align: center;
 }
-  
-  .navbar .user-icon {
-    right: 20px; /* Ajuste para posicionar mais para a direita */
-    width: 50px; /* Tamanho do ícone no navbar */
-    height: 50px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border-radius: 50%;
-    color: white;
-    font-weight: bold;
-    font-size: 24px; /* Ajuste conforme necessário */
-    text-transform: uppercase;
-    cursor: pointer; /* Adiciona o cursor de ponteiro para indicar que é clicável */
-  }
