+++ frontend/src/style/UserBox.css
@@ -1,0 +1,55 @@
+.user-box {
+    position: absolute;
+    top: 80px; /* Ajusta a posição logo abaixo do ícone do usuário */
+    right: 20px; /* Ajuste conforme necessário */
+    background-color: #fff;
+    border: 1px solid #ccc;
+    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+    border-radius: 8px;
+    padding: 20px;
+    text-align: center;
+    z-index: 10; /* Garante que a caixa apareça sobre outros elementos */
+    width: 200px; /* Ajuste a largura conforme necessário */
+  }
+  
+  .user-box-icon {
+    width: 80px; /* Tamanho maior para o ícone dentro da caixa */
+    height: 80px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border-radius: 50%;
+    color: white;
+    font-weight: bold;
+    font-size: 36px;
+    text-transform: uppercase;
+    margin: 0 auto 10px auto; /* Centraliza o ícone e adiciona margem inferior */
+  }
+  
+  .user-box-name {
+    font-size: 18px;
+    font-weight: bold;
+    margin-bottom: 5px;
+  }
+  
+  .user-box-type {
+    font-size: 14px;
+    margin-bottom: 15px;
+  }
+  
+  .user-box-button {
+    display: block;
+    width: 100%;
+    padding: 10px;
+    margin-bottom: 10px;
+    border: 1px solid #ccc;
+    border-radius: 20px;
+    background-color: #fff;
+    color: rgb(0, 0, 0);
+    cursor: pointer;
+  }
+  
+  .user-box-button:hover {
+    background-color: #DB1E2F; /* Cor de fundo ao passar o mouse */
+    color: #fff
+  }
