+++ frontend/src/utils/functions.js
@@ -1,0 +1,11 @@
+
+const getRandomColor = () => {
+    const letters = '0123456789ABCDEF';
+    let color = '#';
+    for (let i = 0; i < 6; i++) {
+      color += letters[Math.floor(Math.random() * 16)];
+    }
+    return color;
+  };
+
+module.exports = {getRandomColor};
